refactor(admin): sidebar restructure + align /admin URL paths
Follow-up on the unified-authentication PR — same admin-UX scope, kept on
the same branch so the navigation reshuffle ships in one reviewable diff.

## Sidebar reshuffle

- **"Infrastructure" section renamed to "PowerDNS"** — operators kept
  asking "infrastructure of what?". The new section heading carries the
  protocol context so the entries underneath can drop their prefixes.
- **Entry rename "PowerDNS servers" → "Servers"** (page heading + browser
  tab title keep the longer "PowerDNS servers" form for clarity in
  isolation).
- **Entry rename "Groups" → "Clusters"** (matching the underlying
  domain — the schema is `pdns_clusters`, the cluster-picker code talks
  about clusters; "Groups" was a UI carry-over).
- **"Request log" moved from "System" into "PowerDNS"** — it's PDNS HTTP
  traffic, not platform audit. System now contains only Settings + Audit
  log.
- `hasInfrastructure` predicate widened to include `canReadAudit` so the
  Request-log entry doesn't disappear behind an empty section header.

## URL alignment (`/admin` paths)

Two paths carried a `pdns-` prefix that the new section heading makes
redundant. Renamed + redirect stubs left at the old paths so bookmarks
and audit-log links keep working:

- `/admin/pdns-clusters` → `/admin/clusters`
- `/admin/pdns-requests` → `/admin/requests`

The other admin paths (`/admin/servers`, `/admin/tsig-keys`,
`/admin/autoprimaries`, `/admin/zone-templates`, `/admin/audit`,
`/admin/settings`, `/admin/users`, `/admin/teams`, `/admin/roles`,
`/admin/authentication`) are already kebab-case plural-noun and don't
need touching. The API surface (`/api/admin/pdns-servers`,
`/api/admin/pdns/clusters`, `/api/admin/pdns/zones`) carries its own
PowerDNS sub-namespace by design — separate, intentional, untouched.

## Page-title alignment

- `/admin/clusters` browser tab + h1 now say "Clusters" (was "Groups").
- `/admin/servers` keeps its "PowerDNS servers" heading + browser title
  (full name on the page itself is clearer than a bare "Servers"
  without the sidebar context).

## Validation gates (all green)

- `act -j static-checks` ✓
- `act -j test` ✓ (893 unit passes)
- `npm run build` ✓ (new routes registered, redirect stubs present)
- `npm run test:integration` ✓ (203 passes — including the renamed
  pdns-clusters integration test file at its new path)
