+++ app/(app)/admin/clusters/new/page.tsx
@@ -22,7 +22,7 @@ export default async function NewClusterPage() {
     <div className="mx-auto max-w-3xl space-y-6">
       <div>
         <Link
-          href="/admin/pdns-clusters"
+          href="/admin/clusters"
           className="text-sm text-[color:var(--color-accent)] hover:underline"
         >
           ← Back to groups
