+++ app/(app)/layout.tsx
@@ -127,9 +127,10 @@ export default async function AppLayout({ children }: Readonly<{ children: React
 
   // Sidebar is grouped by function (Infrastructure / Access / System) rather
   // than one flat "Admin" pile. Each group renders only when the user can see
-  // at least one of its children.
+  // at least one of its children. The PowerDNS section now also houses the
+  // Request log (moved from System), which is `audit.read`-gated.
   const hasInfrastructure =
-    canReadServers || canReadTsig || canManageAutoprimary || canUseTemplates;
+    canReadServers || canReadTsig || canManageAutoprimary || canUseTemplates || canReadAudit;
   const hasAccess = canReadUsers || canReadRoles || canReadTeams || canReadOidc;
   const hasSystem = canReadSettings || canReadAudit;
 
@@ -154,11 +155,9 @@ export default async function AppLayout({ children }: Readonly<{ children: React
         {canReadZones ? <NavLink href="/zones" label="Zones" /> : null}
 
         {hasInfrastructure ? (
-          <NavSection label="Infrastructure">
-            {canReadServers ? (
-              <NavLink nested href="/admin/servers" label="PowerDNS servers" />
-            ) : null}
-            {canReadServers ? <NavLink nested href="/admin/pdns-clusters" label="Groups" /> : null}
+          <NavSection label="PowerDNS">
+            {canReadServers ? <NavLink nested href="/admin/servers" label="Servers" /> : null}
+            {canReadServers ? <NavLink nested href="/admin/clusters" label="Clusters" /> : null}
             {canReadTsig ? <NavLink nested href="/admin/tsig-keys" label="TSIG keys" /> : null}
             {canManageAutoprimary ? (
               <NavLink nested href="/admin/autoprimaries" label="Autoprimaries" />
@@ -166,6 +165,7 @@ export default async function AppLayout({ children }: Readonly<{ children: React
             {canUseTemplates ? (
               <NavLink nested href="/admin/zone-templates" label="Zone templates" />
             ) : null}
+            {canReadAudit ? <NavLink nested href="/admin/requests" label="Request log" /> : null}
           </NavSection>
         ) : null}
 
@@ -184,9 +184,6 @@ export default async function AppLayout({ children }: Readonly<{ children: React
           <NavSection label="System">
             {canReadSettings ? <NavLink nested href="/admin/settings" label="Settings" /> : null}
             {canReadAudit ? <NavLink nested href="/admin/audit" label="Audit log" /> : null}
-            {canReadAudit ? (
-              <NavLink nested href="/admin/pdns-requests" label="Request log" />
-            ) : null}
           </NavSection>
         ) : null}
       </nav>
